+++ board-final-back/src/main/java/com/study/boardfinalback/controller/UserController.java
@@ -3,9 +3,9 @@ package com.study.boardfinalback.controller;
 import com.study.boardfinalback.domain.user.User;
 import com.study.boardfinalback.domain.user.UserRole;
 import com.study.boardfinalback.domain.user.UserStatus;
-import com.study.boardfinalback.dto.LoginDto;
-import com.study.boardfinalback.dto.RegisterDto;
-import com.study.boardfinalback.dto.PasswordChangeDto;
+import com.study.boardfinalback.dto.users.LoginDto;
+import com.study.boardfinalback.dto.users.RegisterDto;
+import com.study.boardfinalback.dto.users.PasswordChangeDto;
 import com.study.boardfinalback.utils.EncryptUtils;
 import com.study.boardfinalback.utils.JwtUtils;
 import com.study.boardfinalback.service.UserService;
